\name{fhesc}
\Rdversion{1.1}
\alias{fhesc}
\docType{data}
\title{
Data from a simple experiment on Human stem cells.
}
\description{
Human stem cells were assayed using Affymetrix 133plus 2 arrays. There
were six arrays, three were biological replicates for undifferentiated
cells, the other three were biological replicates for differentiated cells.
}
\usage{data(fhesc)}
\format{
The data are in the form of an \code{ExpressionSet} instance.
}
\details{

}
\source{
The data were obtained from Dr. M. Tewari.
}
\references{

}
\examples{
data(fhesc)
}
\keyword{datasets}
